requirements.txt
aiohttp==3.9.1
APScheduler==3.10.4
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0
pydantic==2.5.1
notion-client==2.2.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
